requirements.txt
streamlit==1.38.0
langchain-openai==0.1.23
crewai==0.55.2
crewai-tools==0.12.0
agentops==0.3.10
exa-py==1.1.0
firecrawl-py==1.2.3
pysqlite3-binary